4.22 bloglist

+++ part4/blogs/controllers/blogs.js
@@ -1,8 +1,6 @@
 const blogsRouter = require('express').Router()
 const { request } = require('../app')
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -12,11 +10,7 @@ blogsRouter.get('/', async (request, response) => {
 
   blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!decodedToken.id) {    
-      return response.status(401).json({ error: 'token invalid' })  
-    }  
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     const blog = new Blog({
       title: body.title,
@@ -33,6 +27,11 @@ blogsRouter.get('/', async (request, response) => {
   })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({ error: 'not authorized to delete this blog' });
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
   })
